Moved config parsing to module.
